+++ src-gui/src/renderer/components/App.tsx
@@ -15,6 +15,7 @@ import { useEffect } from "react";
 import { setupBackgroundTasks } from "renderer/background";
 import "@fontsource/roboto";
 import FeedbackPage from "./pages/feedback/FeedbackPage";
+import IntroductionModal from "./modal/introduction/IntroductionModal";
 
 const useStyles = makeStyles((theme) => ({
   innerContent: {
@@ -36,6 +37,7 @@ export default function App() {
     <ThemeProvider theme={themes[theme]}>
       <GlobalSnackbarProvider>
         <CssBaseline />
+        <IntroductionModal/>
         <Router>
           <Navigation />
           <InnerContent />