+++ src-gui/src/renderer/components/modal/introduction/slides/Slide03_PrepareSwap.tsx
@@ -1,0 +1,13 @@
+import { Typography } from '@material-ui/core'
+import SlideTemplate from './SlideTemplate'
+import imagePath from 'assets/mockConfigureSwap.svg' 
+
+export default function Slide02_ChooseAMaker(props: slideProps) {
+    return (
+        <SlideTemplate title="Prepare Swap" stepLabel="Step 2" {...props} imagePath={imagePath}>
+            <Typography variant="subtitle1">
+                To initiate a swap, provide a Monero address and optionally a Bitcoin refund address.
+            </Typography>
+        </SlideTemplate>
+    )
+}