+++ src-gui/src/renderer/components/modal/introduction/slides/SlideTypes.ts
@@ -1,0 +1,5 @@
+type slideProps = {
+    handleContinue: () => void
+    handlePrevious: () => void
+    hidePreviousButton?: boolean
+}